Delete maxime-laforet-introduction-to-cssnext.md
We are going to add a language on the website we plan to push the french version (putaindecode.fr -> putaindecode.io/(fr|en)/).

So, since the post has not been published yet, we would like to keep it for our website (ref: https://github.com/putaindecode/organisation/issues/39).

Tell us what you think. I hope it's ok for you.